Invoice/requirements.txt
wheel
tensorflow
wave
langid
pydub
moviepy
eyed3
noisereduce
pyaudioanalysis
pyannote.audio
TTS
spleeter